rssmon/web/requirements.txt
Flask
Flask-SQLAlchemy
feedparser==6.0.2
python-qbittorrent==0.4.2
flask-crontab==0.1.2
urllib3==1.26.3
certifi==2020.12.5
Flask-APScheduler==1.11.0
Flask-Caching==1.10.0
SQLAlchemy==1.4.3
